backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.28.0
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.1

email-validator>=2.1.0
